Get rid of `BlockingTransferData`
Also get rid of one piece of extra debugging
